+++ lib/fst/interface/ifsttable.h
@@ -33,38 +33,38 @@ class IFstTable
   public:
     virtual ~IFstTable() {};
 
-    virtual FstColumnType ColumnType(unsigned int colNr, FstColumnAttribute &columnAttribute, short int &scale,
+    virtual FstColumnType ColumnType(uint32_t colNr, FstColumnAttribute &columnAttribute, short int &scale,
       std::string &annotation, bool &hasAnnotation) = 0;
 
   // Writer interface
-    virtual IStringWriter* GetStringWriter(unsigned int colNr) = 0;
+    virtual IStringWriter* GetStringWriter(uint32_t colNr) = 0;
 
-    virtual int* GetLogicalWriter(unsigned int colNr) = 0;
+    virtual int* GetLogicalWriter(uint32_t colNr) = 0;
 
-    virtual int* GetIntWriter(unsigned int colNr) = 0;
+    virtual int* GetIntWriter(uint32_t colNr) = 0;
 
-    virtual long long* GetInt64Writer(unsigned int colNr) = 0;
+    virtual long long* GetInt64Writer(uint32_t colNr) = 0;
 
-    virtual char* GetByteWriter(unsigned int colNr) = 0;
+    virtual char* GetByteWriter(uint32_t colNr) = 0;
 
-    virtual double* GetDoubleWriter(unsigned int colNr) = 0;
+    virtual double* GetDoubleWriter(uint32_t colNr) = 0;
 
-    virtual IByteBlockColumn* GetByteBlockWriter(unsigned int col_nr) = 0;
+    virtual IByteBlockColumn* GetByteBlockWriter(uint32_t col_nr) = 0;
 
-    virtual IStringWriter* GetLevelWriter(unsigned int colNr) = 0;
+    virtual IStringWriter* GetLevelWriter(uint32_t colNr) = 0;
 
     virtual IStringWriter* GetColNameWriter() = 0;
 
     virtual void GetKeyColumns(int* keyColPos) = 0;
 
-    virtual unsigned int NrOfKeys() = 0;
+    virtual uint32_t NrOfKeys() = 0;
 
-    virtual unsigned int NrOfColumns() = 0;
+    virtual uint32_t NrOfColumns() = 0;
 
-    virtual unsigned long long NrOfRows() = 0;
+    virtual uint64_t NrOfRows() = 0;
 
 	  // Reader interface
-    virtual void InitTable(unsigned int nrOfCols, unsigned long long nrOfRows) = 0;
+    virtual void InitTable(uint32_t nrOfCols, uint64_t nrOfRows) = 0;
 
     virtual IByteBlockColumn* add_byte_block_column(unsigned col_nr) = 0;
 
@@ -85,7 +85,7 @@ class IFstTable
     // use more efficient string container here (e.g. std::vector<string>)
     virtual void SetColNames(IStringArray* col_names) = 0;
 
-    virtual void SetKeyColumns(int* keyColPos, unsigned int nrOfKeys) = 0;
+    virtual void SetKeyColumns(int* keyColPos, uint32_t nrOfKeys) = 0;
 };
 
 #endif  // IFST_TABLE_H
